+++ var/httpd/htdocs/js/Core.Agent.TicketProcess.js
@@ -143,16 +143,6 @@ Core.Agent.TicketProcess = (function (TargetNS) {
                         // initialize ajax dnd upload
                         Core.UI.InitAjaxDnDUpload();
 
-                        // move help triggers into field rows for dynamic fields
-                        $('.Row > .FieldHelpContainer').each(function () {
-                            if (!$(this).next('label').find('.Marker').length) {
-                                $(this).prependTo($(this).next('label'));
-                            }
-                            else {
-                                $(this).insertAfter($(this).next('label').find('.Marker'));
-                            }
-                        });
-
                         // Initially display dynamic fields with TreeMode = 1 correctly
                         Core.UI.TreeSelection.InitDynamicFieldTreeViewRestore();